refactor: Extract CDN functionality into dedicated module
