+++ src/commands/help.ts
@@ -6,9 +6,22 @@ module.exports = {
         .setDescription('List of custom commands available for all Commanders'),
     async execute(interaction: CommandInteraction) {
         await interaction.reply({
-            content: `CUSTOM COMMANDS \n 
+            content: `====================\n**SLASH COMMANDS**\n====================\n
 ➜ **/gacha <type> <game>** : Gamba with rapi bot, you can do singles or multis, works with NIKKE
 ➜ **/lucky** : How lucky are you today?
+➜ **/rapiball** : Get an eight ball of Rapi's wisdom
+➜ **/compositions** : Get help for NIKKE Team Compositions
+➜ **/daily time <game>** : Gets the daily reset for the specified game
+➜ **/rules** : Get the rules for the server
+➜ **/relics** : Get lost relics guide in NIKKE
+➜ **stream** : Reserved for El Shafto
+
+
+====================
+**CHAT COMMANDS**
+====================
+
+
 ➜ **damn train** : Fuck Thomas The Train
 ➜ **damn gravedigger** : Time for OSU!
 ➜ **good girl** : say thanks to the best girl & bot in this server 
@@ -35,6 +48,8 @@ module.exports = {
 ➜ **is it over?** : Yes it is, Commander.
 ➜ **99%** : Here's some advice for you.
 ➜ **belorta...** : CURSE OF BELORTA.
+➜ **get dat nikke** : Send Rapi to fight your rival.
+
 `,
             ephemeral: true
         });
