+++ src/commands/utils/gachaPuller.ts
@@ -1,5 +1,5 @@
 import { ListObjectsV2Command } from "@aws-sdk/client-s3";
-import { s3Client } from '../../discord';
+import { s3Client } from '../../utils/cdn';
 import { CONSTANTS } from './gachaConstants';
 import { GachaGameConfig, PullResult } from './gachaTypes';
 import { NikkeUtil } from './nikkeUtil';
