+++ src/discord.ts
@@ -29,11 +29,10 @@ import 'moment-timezone';
 import * as util from "./utils/util";
 import { VoiceConnectionData } from "./utils/interfaces/VoiceConnectionData.interface";
 import { getCCPMessage } from "./utils/constants/messages";
-
-import { S3, S3ClientConfig, ListObjectsV2Command } from "@aws-sdk/client-s3";
 import { SensitiveTerm } from "./utils/interfaces/SensitiveTerm.interface";
 import { Command, MessageCommand } from './utils/interfaces/Command.interface';
 import { CustomClient } from "./utils/interfaces/CustomClient.interface";
+import { getRandomCdnMediaUrl } from "./utils/cdn/mediaManager";
 
 // Destructure only the necessary functions from util
 const {
@@ -52,21 +51,8 @@ const {
     isMessageCommand
 } = util;
 
-//TODO: Extract into common utils S3 client
-const s3ClientConfig: S3ClientConfig = {
-    forcePathStyle: false,
-    endpoint: "https://sfo3.digitaloceanspaces.com",
-    region: "us-east-1",
-    credentials: {
-      accessKeyId: process.env.S3ACCESSKEYID as string,
-      secretAccessKey: process.env.S3SECRETKEY as string
-    }
-};
-const s3Client = new S3(s3ClientConfig);
-
 const DISCORD_TOKEN = process.env.WAIFUTOKEN as string;
 const CLIENT_ID = process.env.CLIENTID as string;
-const S3_BUCKET = process.env.S3BUCKET as string;
 const RADIO_FOLDER_PATH = './src/radio';
 const PRE = "/";
 const NIKKE_RESET_START_TIME = moment.tz({ hour: 20, minute: 0, second: 0, millisecond: 0 }, 'UTC');
@@ -82,7 +68,7 @@ const ASSET_PATHS = {
     }
 } as const;
 
-// CDN URLs
+// Asset URLs
 const RAPI_BOT_THUMBNAIL_URL = `${cdnDomainUrl}/${ASSET_PATHS.THUMBNAIL}`;
 const GFL2_LOGO_URL = `${cdnDomainUrl}/${ASSET_PATHS.LOGOS.GFL2}`;
 const NIKKE_LOGO_URL = `${cdnDomainUrl}/${ASSET_PATHS.LOGOS.NIKKE}`;
@@ -107,8 +93,6 @@ const bot: CustomClient = new Client({
 
 bot.commands = new Collection();
 const commands: Array<object> = [];
-
-
 
 // Extend Client to include commands
 declare module 'discord.js' {
@@ -1872,135 +1856,5 @@ async function initDiscordBot() {
 
 export {
     initDiscordBot,
-    bot as getDiscordBot, s3Client,
+    bot as getDiscordBot,
 };
-
-// Track recently sent media keys per guild and prefix
-const recentlySentMediaKeys: Map<string, Map<string, string[]>> = new Map();
-
-/**
- * Retrieves and filters media from the CDN based on specified criteria, with guild-based tracking 
- * to prevent recent repeats within each server.
- * 
- * @param prefix - The bucket prefix path (e.g., "commands/leadership/")
- * @param guildId - Discord guild ID for tracking media per server
- * @param options - Optional configuration object
- * @param options.maxSizeMB - Maximum file size in MB (default: 100)
- * @param options.extensions - Array of allowed file extensions (default: ['.mp4'])
- * @param options.trackLast - Number of media to track for preventing repeats (default: 5)
- * @returns Promise<string> A fully qualified CDN URL (e.g., 'https://rapi-bot.sfo3.cdn.digitaloceanspaces.com/commands/leadership/video.mp4')
- * 
- * @throws {Error} When prefix or guildId is missing/invalid
- * @throws {Error} When no valid media files are found
- * @throws {Error} When S3 operations fail
- * 
- * @example
- * // Basic usage with default options
- * const url = await getRandomCdnMediaUrl("commands/leadership/", "123456789");
- * 
- * // With custom options
- * const url = await getRandomCdnMediaUrl("commands/videos/", "123456789", {
- *   maxSizeMB: 50,
- *   extensions: DEFAULT_IMAGE_EXTENSIONS,
- *   trackLast: 3
- * });
- */
-async function getRandomCdnMediaUrl(
-    prefix: string,
-    guildId: string,
-    options: {
-        maxSizeMB?: number;
-        extensions?: string[];
-        trackLast?: number;
-    } = {}
-): Promise<string> {
-    // Validate required parameters
-    if (!prefix || typeof prefix !== 'string') {
-        console.error('Invalid prefix provided:', prefix);
-        throw new Error('Valid prefix path is required');
-    }
-
-    if (!guildId || typeof guildId !== 'string') {
-        console.error('Invalid guildId provided:', guildId);
-        throw new Error('Valid guild ID is required');
-    }
-
-    const maxSizeMB = options.maxSizeMB ?? 100;
-    const extensions = options.extensions ?? [...DEFAULT_IMAGE_EXTENSIONS, ...DEFAULT_VIDEO_EXTENSIONS];
-    const trackLast = options.trackLast ?? 5;
-
-    try {
-        console.log(`Fetching media for prefix: ${prefix} in guild: ${guildId}`);
-
-        // List objects in the specified folder
-        const response = await s3Client.send(new ListObjectsV2Command({
-            Bucket: S3_BUCKET,
-            Prefix: prefix,
-        }));
-
-        if (!response.Contents || response.Contents.length === 0) {
-            console.error(`No media files found in CDN under prefix: ${prefix}`);
-            throw new Error('No media files available');
-        }
-
-        // Filter media files based on size and extension
-        const mediaKeys = response.Contents
-            .filter(obj => {
-                const sizeInMB = (obj.Size || 0) / (1024 * 1024);
-                return sizeInMB <= maxSizeMB;
-            })
-            .map(obj => obj.Key)
-            .filter(key => key && extensions.some(ext => key.toLowerCase().endsWith(ext)));
-
-        if (mediaKeys.length === 0) {
-            console.error(`No valid media files found under ${maxSizeMB}MB with extensions: ${extensions.join(', ')}`);
-            throw new Error('No suitable media files available');
-        }
-
-        // Initialize guild tracking if not exists
-        if (!recentlySentMediaKeys.has(guildId)) {
-            console.log(`Initializing tracking for new guild: ${guildId}`);
-            recentlySentMediaKeys.set(guildId, new Map());
-        }
-
-        // Initialize prefix tracking for this guild if not exists
-        const guildTracking = recentlySentMediaKeys.get(guildId)!;
-        if (!guildTracking.has(prefix)) {
-            console.log(`Initializing tracking for prefix: ${prefix} in guild: ${guildId}`);
-            guildTracking.set(prefix, []);
-        }
-
-        const recentKeys = guildTracking.get(prefix)!;
-        console.log(`Currently tracking ${recentKeys.length} keys for ${prefix} in guild: ${guildId}`);
-
-        // Filter out recently sent keys
-        const availableKeys = mediaKeys.filter(key => key && !recentKeys.includes(key));
-
-        // If all keys have been recently used, reset tracking and use all keys
-        if (availableKeys.length === 0) {
-            console.log(`All media for ${prefix} have been recently used in guild: ${guildId}. Resetting tracking.`);
-            guildTracking.set(prefix, []);
-            const randomCdnMediaKey = mediaKeys[Math.floor(Math.random() * mediaKeys.length)]!;
-            guildTracking.get(prefix)!.push(randomCdnMediaKey);
-            return `${cdnDomainUrl}/${randomCdnMediaKey}`;
-        }
-
-        // Select a random key from available ones
-        const randomCdnMediaKey = availableKeys[Math.floor(Math.random() * availableKeys.length)]!;
-        
-        // Add to tracking
-        recentKeys.push(randomCdnMediaKey);
-        
-        // Keep only the last trackLast number of keys
-        if (recentKeys.length > trackLast) {
-            recentKeys.shift();
-        }
-
-        console.log(`Selected random key: ${randomCdnMediaKey} for guild: ${guildId}`);
-        return `${cdnDomainUrl}/${randomCdnMediaKey}`;
-
-    } catch (error) {
-        console.error(`Error retrieving CDN media for guild ${guildId}:`, error);
-        throw error instanceof Error ? error : new Error('Error retrieving CDN media');
-    }
-}
