+++ src/utils/interfaces/MediaTracking.interface.ts
@@ -1,0 +1,5 @@
+export interface MediaTracking {
+    guildId: string;
+    prefix: string;
+    recentKeys: string[];
+}
